requirements.txt
python-telegram-bot>=21.0
matplotlib>=3.7
openpyxl>=3.1
